Add a test to ensure file headers are formatted properly
I am trying to solve https://github.com/genonbeta/TrebleShot/issues/105 and I just realized in the test cases, NanoHTTPD developers doesn't use the formatting method that is suggested on the forums to get file upload names. They are using the Apache Commons libraries to do so and the issue is not visible for that reason. Now I learned that I am going to integrate the Apache classes for more consistency like it happens in the test cases. Also, I am going to convert Response class to an interface so that different ways of responses can be created.