Add ServerExecutor class to execute the server instance
Also, I made the errrors available to be gathered by the calling thread (sent by the runner thread). This is important because if the server fails to start and there is a request right after the start is called, it would also fail. The next good idea would be to make the start and stop calls non-blocking where it is not important if the server starts or stops as those two methods return.